follow dumh and smart component pattern
